Refactoring & clean-up of anonymous messaging
Renamed MessageController to ParticipantCommunicationController and moved to controller folder.  Renamed MessageFormType to AnonymousMessageFormType. And clean-up of translations regarding anonymous messaging. See #222
