+++ src/Intracto/SecretSantaBundle/Form/Type/AnonymousMessageFormType.php
@@ -7,23 +7,19 @@ use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Validator\Constraints\NotBlank;
-use Symfony\Component\Validator\Constraints\Length;
 
-class MessageFormType extends AbstractType
+class AnonymousMessageFormType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
              ->add('message', TextareaType::class, [
-                'label' => 'entry_show_valid.send_message.message_label',
+                'label' => 'entry_show_valid.anonymous_message.message_label',
                 'constraints' => [
                     new NotBlank(),
-                    new Length([
-                        'min' => 10,
-                    ]),
                 ],
                 'attr' => [
-                    'placeholder' => 'entry_show_valid.send_message.message_placeholder',
+                    'placeholder' => 'entry_show_valid.anonymous_message.message_placeholder',
                 ],
              ])
              ->add('recipient', HiddenType::class)
