+++ src/Intracto/SecretSantaBundle/Mailer/MailerService.php
@@ -430,32 +430,22 @@ class MailerService
     /**
      * @param $recipient
      * @param $message
-     *
-     * @return bool
      */
     public function sendAnonymousMessage($recipient, $message)
     {
-        $addressee = $this->em->getRepository('IntractoSecretSantaBundle:Entry')->find($recipient);
-
-        if (count($addressee) == 0) {
-            return false;
-        }
-
-        $pool = $this->em->getRepository('IntractoSecretSantaBundle:Pool')->find($addressee->getPool());
-
-        $this->translator->setLocale($pool->getLocale());
+        $this->translator->setLocale($recipient->getPool()->getLocale());
 
         $message = \Swift_Message::newInstance()
             ->setSubject($this->translator->trans('emails-message.subject'))
             ->setFrom($this->noreplyEmail, $this->translator->trans('emails-base_email.sender'))
-            ->setTo($addressee->getEmail())
+            ->setTo($recipient->getEmail())
             ->setBody(
                 $this->templating->render(
                     'IntractoSecretSantaBundle:Emails:anonymousMessage.html.twig',
                     [
-                        'name' => $addressee->getName(),
+                        'name' => $recipient->getName(),
                         'message' => $message,
-                        'entry' => $addressee,
+                        'entry' => $recipient,
                     ]
                 ),
                 'text/html'
@@ -464,15 +454,13 @@ class MailerService
                 $this->templating->render(
                     'IntractoSecretSantaBundle:Emails:anonymousMessage.txt.twig',
                     [
-                        'name' => $addressee->getName(),
+                        'name' => $recipient->getName(),
                         'message' => $message,
-                        'entry' => $addressee,
+                        'entry' => $recipient,
                     ]
                 ),
                 'text/plain'
             );
         $this->mailer->send($message);
-
-        return true;
     }
 }
