+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.fr.yml
@@ -223,19 +223,19 @@ entry_show_valid:
         update_item: Actualiser cet object
         remove_item: Supprimer
 
-#    send_message:
+#    anonymous_message:
 #        title:
 #        message_label:
 #        message_placeholder:
 #        warning_anonymous:
 #        send_button: Verstuur je bericht!
 
-#MessageController
+# ParticipantCommunicationController
 #send_message:
 #    feedback:
-#              success:
-#              error:
-#              error_in_form:
+#        success:
+#        error:
+#        error_in_form:
 
 # Pool/manage/base.html.twig
 pool_manage_base:
@@ -489,13 +489,13 @@ emails-wishlistChanged:
     btn:
         check_wishlist: Découvrir la liste de souhaits
 
-#Emails/anonymousMessage.{html,txt}.twig
+# Emails/anonymousMessage.{html,txt}.twig
 #emails-message:
-#    subject:
+#    subject: ?
 #    message:
-#      html: >
-#      txt: >
-#    update_wishlist:
+#      html: ?
+#      txt: ?
+#    update_wishlist: ?
 
 flashes:
     # SendPartyUpdateController
