+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.nl.yml
@@ -265,19 +265,19 @@ entry_show_valid:
         update_item: Update dit item
         remove_item: Verwijder
 
-    send_message:
+    anonymous_message:
         title: Stuur een anoniem bericht naar %name%
         message_label: Jouw bericht
         message_placeholder: Jouw bericht
         warning_anonymous: Denk eraan dat dit bericht anoniem wordt verstuurd en je niet mag verklappen wie je bent! Je maatje kan niet antwoorden op je bericht, wees concreet zodat hij/zij zijn/haar wishlist kan updaten.
         send_button: Verstuur je bericht!
 
-#MessageController
+# ParticipantCommunicationController
 send_message:
     feedback:
-              success: Je email is verzonden! Onze kaboutertjes zullen hem bezorgen.
-              error: Helaas kon je bericht niet verzonden worden. Probeer later opnieuw!
-              error_in_form: Er is een fout in je email! Controleer alles en probeer opnieuw!
+        success: Je email is verzonden! Onze kaboutertjes zullen hem bezorgen.
+        error: Helaas kon je bericht niet verzonden worden. Probeer later opnieuw!
+        error_in_form: Er is een fout in je email! Controleer alles en probeer opnieuw!
 
 # Pool/manage/base.html.twig
 pool_manage_base:
@@ -549,7 +549,7 @@ emails-wishlistChanged:
     btn:
         check_wishlist: Bekijk verlanglijst
 
-#Emails/anonymousMessage.{html,txt}.twig
+# Emails/anonymousMessage.{html,txt}.twig
 emails-message:
     subject: Je Secret Santa heeft je een bericht gestuurd!
     message:
@@ -567,7 +567,6 @@ emails-message:
           Je Secret Santa heeft je een bericht gestuurd, je kan het hier lezen:
 
           %message%
-
 
           Klik op de link hier beneden om je lijst te bekijken, en zaken toe te voegen aan je wishlist!
     update_wishlist: Pas je verlanglijst aan
