+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.pl.yml
@@ -176,19 +176,19 @@ entry_show_valid:
 #        update_item: ?
         remove_item: Usuń pozycję
 
-#    send_message:
+#    anonymous_message:
 #        title:
 #        message_label:
 #        message_placeholder:
 #        warning_anonymous:
 #        send_button: Verstuur je bericht!
 
-#MessageController
+# ParticipantCommunicationController
 #send_message:
 #    feedback:
-#              success: Je email is verzonden! Onze kaboutertjes zullen hem bezorgen.
-#              error: Helaas kon je bericht niet verzonden worden. Probeer later opnieuw!
-#              error_in_form: Er is een fout in je email! Controleer alles en probeer opnieuw!
+#        success:
+#        error:
+#        error_in_form:
 
 # Pool/manage/base.html.twig
 pool_manage_base:
@@ -378,13 +378,13 @@ emails-wishlistChanged:
     btn:
         check_wishlist: Sprawdź listę życzeń
 
-#Emails/anonymousMessage.{html,txt}.twig
+# Emails/anonymousMessage.{html,txt}.twig
 #emails-message:
-#    subject:
+#    subject: ?
 #    message:
-#      html: >
-#      txt: >
-#    update_wishlist:
+#      html: ?
+#      txt: ?
+#    update_wishlist: ?
 
 flashes:
     # SendPartyUpdateController
