+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.zh-Hans.yml
@@ -189,19 +189,19 @@ entry_show_valid:
 #        update_item: ?
         remove_item: 删除这个物品
 
-#    send_message:
+#    anonymous_message:
 #        title:
 #        message_label:
 #        message_placeholder:
 #        warning_anonymous:
 #        send_button: Verstuur je bericht!
 
-#MessageController
+# ParticipantCommunicationController
 #send_message:
 #    feedback:
-#              success:
-#              error:
-#              error_in_form:
+#        success:
+#        error:
+#        error_in_form:
 
 # Pool/manage/base.html.twig
 pool_manage_base:
@@ -411,13 +411,13 @@ emails-wishlistChanged:
     btn:
         check_wishlist: 查看愿望清单
 
-#Emails/anonymousMessage.{html,txt}.twig
+# Emails/anonymousMessage.{html,txt}.twig
 #emails-message:
-#    subject:
+#    subject: ?
 #    message:
-#      html: >
-#      txt: >
-#    update_wishlist:
+#      html: ?
+#      txt: ?
+#    update_wishlist:?
 
 flashes:
     # SendPartyUpdateController
