+++ src/Intracto/SecretSantaBundle/Resources/views/Emails/anonymousMessage.txt.twig
@@ -1,3 +1,2 @@
-{% set emessage = message | escape %}
-{{ 'emails-message.message.txt'|trans({"%name%": entry.name , "%message%" : emessage })|raw }}<br>
+{{ 'emails-message.message.txt'|trans({"%name%": entry.name , "%message%" : message | escape })|raw }}<br>
 {{ url("entry_view", { "url": entry.url, "_locale": entry.pool.locale }) }}
